+++ framework/model/src/main/java/is/codion/framework/model/DefaultEntitySearchModel.java
@@ -28,6 +28,8 @@ import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.domain.entity.Entity;
 import is.codion.framework.domain.entity.EntityDefinition;
 import is.codion.framework.domain.entity.EntityType;
+import is.codion.framework.domain.entity.OrderBy;
+import is.codion.framework.domain.entity.OrderBy.OrderByColumn;
 import is.codion.framework.domain.entity.attribute.Attribute;
 import is.codion.framework.domain.entity.attribute.Column;
 import is.codion.framework.domain.entity.condition.Condition;
@@ -44,8 +46,7 @@ import static is.codion.framework.domain.entity.condition.Condition.and;
 import static is.codion.framework.domain.entity.condition.Condition.or;
 import static java.util.Collections.*;
 import static java.util.Objects.requireNonNull;
-import static java.util.stream.Collectors.joining;
-import static java.util.stream.Collectors.toMap;
+import static java.util.stream.Collectors.*;
 
 final class DefaultEntitySearchModel implements EntitySearchModel {
 
@@ -61,6 +62,7 @@ final class DefaultEntitySearchModel implements EntitySearchModel {
 	private final EntityType entityType;
 	private final Collection<Column<String>> columns;
 	private final Collection<Attribute<?>> attributes;
+	private final OrderBy orderBy;
 	private final DefaultSelection selection = new DefaultSelection();
 	private final EntityConnectionProvider connectionProvider;
 	private final Map<Column<String>, Settings> settings;
@@ -85,6 +87,7 @@ final class DefaultEntitySearchModel implements EntitySearchModel {
 		this.separator.set(builder.separator);
 		this.columns = unmodifiableCollection(builder.columns);
 		this.attributes = builder.attributes;
+		this.orderBy = builder.orderBy;
 		this.settings = unmodifiableMap(columns.stream()
 						.collect(toMap(Function.identity(), column -> new DefaultSettings())));
 		this.stringFunction.set(builder.stringFunction);
@@ -198,6 +201,7 @@ final class DefaultEntitySearchModel implements EntitySearchModel {
 		return Select.where(createCombinedCondition(conditions))
 						.attributes(attributes)
 						.limit(limit.get())
+						.orderBy(orderBy)
 						.build();
 	}
 
@@ -328,11 +332,14 @@ final class DefaultEntitySearchModel implements EntitySearchModel {
 		private boolean singleSelection = false;
 		private String separator = DEFAULT_SEPARATOR;
 		private Integer limit = DEFAULT_LIMIT.get();
+		private OrderBy orderBy;
 
 		DefaultBuilder(EntityType entityType, EntityConnectionProvider connectionProvider) {
 			this.entityType = requireNonNull(entityType);
 			this.connectionProvider = requireNonNull(connectionProvider);
-			this.columns = connectionProvider.entities().definition(entityType).columns().searchable();
+			EntityDefinition definition = connectionProvider.entities().definition(entityType);
+			this.columns = definition.columns().searchable();
+			this.orderBy = definition.orderBy().orElse(null);
 		}
 
 		@Override
@@ -349,6 +356,15 @@ final class DefaultEntitySearchModel implements EntitySearchModel {
 		public Builder attributes(Collection<Attribute<?>> attributes) {
 			validateAttributes(requireNonNull(attributes));
 			this.attributes = attributes;
+			return this;
+		}
+
+		@Override
+		public Builder orderBy(OrderBy orderBy) {
+			validateAttributes(requireNonNull(orderBy).orderByColumns().stream()
+							.map(OrderByColumn::column)
+							.collect(toList()));
+			this.orderBy = orderBy;
 			return this;
 		}