+++ changelog.md
@@ -4,6 +4,7 @@ Codion Change Log
 ## 0.18.25
 ## is.codion.framework.model
 - EntitySearchModel.Builder.attributes() added.
+- EntitySearchModel.Builder.orderBy() added.
 
 ## 0.18.24
 ## is.codion.common.core